requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.10.0,<3.0.0
python-multipart>=0.0.18,<1.0.0

# MongoDB
motor>=3.3.0,<4.0.0          # async driver (FastAPI routes)
pymongo>=4.6.0,<5.0.0        # sync driver (agent tools) — installed as motor dependency

# LlamaIndex agent orchestration
llama-index-core>=0.12.0,<1.0.0
llama-index-llms-openai>=0.4.0,<1.0.0
llama-index-agent-openai>=0.4.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# OpenAI SDK (direct — used by RAG service)
openai>=1.30.0,<2.0.0

# Pinecone (RAG vector store)
pinecone>=5.0.0,<6.0.0
